backend/requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
bcrypt==3.2.0
blinker==1.7.0
Brlapi==0.8.3
certifi==2020.6.20
channels==4.0.0
chardet==4.0.0
click==8.1.7
cloudinary==1.36.0
colorama==0.4.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.8.0rc2
Django==3.2.12
django-cloudinary-storage==0.3.0
django-cors-headers==3.5.0
django-nextjs==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dnspython==2.6.1
duplicity==0.8.21
exceptiongroup==1.2.0
fastapi==0.109.2
fasteners==0.14.1
filelock==3.13.1
future==0.18.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
itypes==1.2.0
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.4
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mutagen==1.45.1
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
passlib==1.7.4
pexpect==4.8.0
Pillow==9.0.1
pipenv==11.9.0
platformdirs==4.1.0
protobuf==3.12.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.6.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.4.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.5.3
SQLAlchemy==2.0.27
sqlparse==0.2.4
ssh-import-id==5.11
starlette==0.36.3
stripe==2.58.0
systemd-python==234
twilio==8.11.1
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.27.1
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
wadllib==1.3.6
websockets==12.0
Werkzeug==3.0.1
whitenoise==6.6.0
xdg==5
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
